+++ backend/srcs/src/auth/auth.module.ts
@@ -9,6 +9,7 @@ import {
 import { AuthController } from './auth.controller';
 import { UserService } from '../user/user.service';
 import { FakeAuthService } from './fake.auth.service';
+import { ProfileImagesService, RandomStringService } from 'src/profile_images/profile_images.service';
 
 @Module({
   imports: [
@@ -29,6 +30,8 @@ import { FakeAuthService } from './fake.auth.service';
     FortyTwoStrategy,
     JwtStrategy,
     UserService,
+    ProfileImagesService,
+    RandomStringService,
     FakeAuthService,
     RefreshTokenStrategy
   ],