+++ backend/srcs/src/profile_images/profile_images.module.ts
@@ -1,0 +1,9 @@
+import { Module } from '@nestjs/common';
+import { ProfileImagesService, RandomStringService } from './profile_images.service';
+import { ProfileImagesController } from './profile_images.controller';
+
+@Module({
+  providers: [ProfileImagesService, RandomStringService],
+  controllers: [ProfileImagesController]
+})
+export class ProfileImagesModule {}